+++ nodemon.json
@@ -1,3 +1,3 @@
 {
-  "ignore": ["node_modules", "models/contacts.json"]
+  "ignore": ["node_modules"]
 }
