+++ routes/api/contacts.js
@@ -8,9 +8,9 @@ const { schemas } = require("../../models/contact");
 
 const router = express.Router();
 
-router.get('/', ctrl.listContacts);
+router.get('/', ctrl.getAllContacts);
 
-router.get('/:contactId', isValidId, ctrl.getById);
+router.get('/:contactId', isValidId, ctrl.getContactById);
 
 router.post("/", validateBody(schemas.schema), ctrl.addContact);
 
@@ -18,6 +18,6 @@ router.delete('/:contactId', isValidId, ctrl.removeContact);
 
 router.put('/:contactId', isValidId, validateBody(schemas.schema), ctrl.updateContact);
 
-router.patch('/:contactId/favorite', isValidId, validateBody(schemas.updateFavoriteSchema), ctrl.updateFavorite);
+router.patch('/:contactId/favorite', isValidId, validateBody(schemas.updateFavoriteSchema), ctrl.updateFavoriteContact);
 
 module.exports = router;
